rename jupyterAvailability.ts to jupyterExecution.ts
